// src/enums/calculator.ts
export enum CalculatorButton {
  SEVEN = '7',
  EIGHT = '8',
  NINE = '9',
  DELETE = 'DEL',
  FOUR = '4',
  FIVE = '5',
  SIX = '6',
  ADD = '+',
  ONE = '1',
  TWO = '2',
  THREE = '3',
  SUBTRACT = '-',
  DOT = '.',
  ZERO = '0',
  DIVIDE = '/',
  MULTIPLY = '*',
  RESET = 'RESET',
  EQUAL = '=',
}